Challenge requisites fullfilled, looking out for posible fixes

+++ routes/products/RouteProducts.js
@@ -1,5 +1,6 @@
 const express = require("express");
 const { Router } = express;
+const admin = require("../../controllers/adminController");
 
 const productosRouter = new Router();
 
@@ -9,7 +10,7 @@ productosRouter.get("/", async (req, res) => {
   try {
     const prod = await prods.getAll();
     // res.render("pages/allProducts", { prod });
-    res.send(prod);
+    res.json(prod);
   } catch (error) {
     console.log(error);
   }
@@ -18,38 +19,50 @@ productosRouter.get("/:id", async (req, res) => {
   try {
     const prod = await prods.getById(req.params.id);
     // res.render("pages/products", prod);
-    res.send(prod);
+    res.json(prod);
   } catch (error) {
     console.log(error);
   }
 });
 
 productosRouter.post("/", async (req, res) => {
-  try {
-    const prod = await prods.save(req.body);
-    // res.render("pages/products", prod);
-    res.send(prod);
-  } catch (error) {
-    console.log(error);
+  if (admin) {
+    try {
+      const prod = await prods.save(req.body);
+      // res.render("pages/products", prod);
+      res.json(prod);
+    } catch (error) {
+      console.log(error);
+    }
+  } else {
+    res.json({ error: -1, descripcion: "route '/' method 'POST' unauthorized" });
   }
 });
 
 productosRouter.put("/:id", async (req, res) => {
-  try {
-    const prod = await prods.edit(req.params.id, req.body);
-    // res.render("pages/products", prod);
-    res.send(prod);
-  } catch (error) {
-    console.log(error);
+  if (admin) {
+    try {
+      const prod = await prods.edit(req.params.id, req.body);
+      // res.render("pages/products", prod);
+      res.json(prod);
+    } catch (error) {
+      console.log(error);
+    }
+  } else {
+    res.json({ error: -1, descripcion: "route '/req.params.id' method 'PUT' unauthorized" });
   }
 });
 
 productosRouter.delete("/:id", async (req, res) => {
-  try {
-    const prod = await prods.deleteById(req.params.id);
-    res.send(prod);
-  } catch (error) {
-    console.log(error);
+  if (admin) {
+    try {
+      const prod = await prods.deleteById(req.params.id);
+      res.json(prod);
+    } catch (error) {
+      console.log(error);
+    }
+  } else {
+    res.json({ error: -1, descripcion: `route '/${req.params.id}' method 'DELETE' unauthorized` });
   }
 });
 module.exports = productosRouter;
